+++ backend/src/main/java/com/carffeine/carffeine/member/controller/dto/JwtResponse.java
@@ -1,0 +1,4 @@
+package com.carffeine.carffeine.member.controller.dto;
+
+public record JwtResponse(String jwtToken) {
+}
