+++ backend/src/main/java/com/carffeine/carffeine/member/controller/dto/MemberResponse.java
@@ -1,0 +1,4 @@
+package com.carffeine.carffeine.member.controller.dto;
+
+public record MemberResponse(Long id, String email) {
+}
