+++ backend/src/main/java/com/carffeine/carffeine/member/domain/TokenProvider.java
@@ -1,0 +1,10 @@
+package com.carffeine.carffeine.member.domain;
+
+public interface TokenProvider {
+
+    String createJwt(Long id);
+
+    boolean isExpired(String token);
+
+    Long extract(String token);
+}
