--- backend/src/main/java/com/carffeine/carffeine/station/service/auth/MemberService.java
@@ -1,16 +1,0 @@
-package com.carffeine.carffeine.station.service.auth;
-
-import com.carffeine.carffeine.station.domain.jwt.Jwt;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class MemberService {
-
-    private final Jwt jwt;
-
-    public String createJwt(Long memberId) {
-        return jwt.createJwt(memberId);
-    }
-}
